Define classes and modules under NArray class
This defines all classes and modules under the NArray class (other than
the NArray class itself, of course).  This is done to reduce the
potential for namespace conflicts at the top level.  This change has
been done in a backwards compatible way so it should not break existing
code.  New code that wants to use the namespaced classes and modules
must do:

  require 'narray/narray'

That will define the NArray class at the top level and the following
classes and modules under the NArray class:

  NArray::NMath
  NArray::NVector
  NArray::NMatrix
  NArray::NMatrixLU
  NArray::NArrayScalar

Existing code that does:

  require 'narray'

will work as before because that loads the namespaced classes and
modules and then creates top level aliases for them.